Better buffer serialization
This eliminates the need for the hack in the broadcast helper that caused it to have side-effects.

There is a performance hit to this (buffer is encoded then decoded and encoded again...) but it's preferable to having side effects or monkey-patching Buffer.prototype.toJSON. Users can avoid this performance hit by making sure to pass dsteem's HexBuffer objects.